c float

+++ term1/block3/hypothenuse.py
@@ -1,5 +1,5 @@
 import math
 a=float(input("введите длину первого катета:"))
 b=float(input("введите длину второго катета:"))
-с=math.sqrt(a**2+b**2)
+с=float(math.sqrt(a**2+b**2))
 print (c)
